# process replay: only enter prefix when interacting with process