Only enter prefix when ready to send/receive msgs, not simply adding messages to queue (list)

---

benchmark command:

```bash
./run_process_on_route.py d9b97c1d3b8c39b2/00000160--6ce3dd7c70/1 selfdrived controlsd radard plannerd calibrationd dmonitoringd locationd paramsd lagd ubloxd torqued card
```

On master:

```
Total run_step times:
  selfdrived: 5.788743521086872s
  card: 1.0108267199248075s
  locationd: 0.4765220759436488s
  controlsd: 2.740807613823563s
  radard: 0.3774205078370869s
  plannerd: 0.9272696250118315s
  dmonitoringd: 0.545840504579246s
  calibrationd: 0.22642212361097336s
  paramsd: 0.2980232019908726s
  lagd: 0.5793828601017594s
  torqued: 0.6220222664996982s
Total run_step time: 13.59s
```

This PR:
```
Total run_step times:
  selfdrived: 5.424799048807472s
  card: 0.919613890349865s
  locationd: 0.3173013483174145s
  controlsd: 2.510532932356s
  radard: 0.28098296048119664s
  plannerd: 0.586313575040549s
  dmonitoringd: 0.3731035953387618s
  calibrationd: 0.17976072849705815s
  paramsd: 0.23841818049550056s
  lagd: 0.3673454560339451s
  torqued: 0.4409753093495965s
Total run_step time: 11.64s
```